1.결제 관련 오류 수정 2.잔금 결제 상품 환불 시 잔금, 예약금 한번에 환불

+++ src/main/java/com/soloProject/myTrip/entity/Payment.java
@@ -29,7 +29,7 @@ public class Payment extends BaseTimeEntity {
    private String merchantUid; // 주문 번호
 
    @Column(name = "amount", nullable = false)
-   private BigDecimal amount; // 결제된 금액
+   private Integer amount; // 결제된 금액
 
    private String paymentKey;
